Refactor usecases into a separate submodule
